include/gitlab: Make urlencode function more extensible
Because _gitlab_urlencode() is a simple string replacement, it is
not very extensible and does not handle more than one replacement.

This commit rewrites _gitlab_urlencode() to use an associative
array to store replacement patterns, allowing the function to be
easily extended. For testing, this commit also adds a pattern to
escape comma (',') characters in URLs.